outsource ACRClient from main script
